Remove component collapsing and read-only mode. Remove StackedItemComponent prior to EditorComponent
