+++ src/app/editor/components/AbstractComponent.ts
@@ -1,9 +1,14 @@
-import { Input } from '@angular/core';
+import { Input,ComponentResolver,ViewContainerRef } from '@angular/core';
 import { EditorComponent } from './EditorComponent';
 
 export abstract class AbstractComponent<T> extends EditorComponent {
   @Input()
   public model: T;
+
+  constructor(componentResolver:ComponentResolver,
+              viewContainerRef:ViewContainerRef) {
+    super(componentResolver, viewContainerRef);
+  }
   
   public onReset(): void {
     for (let prop in this.model) { delete this.model[prop] }
