+++ src/app/editor/components/CardHeaderComponent.html
@@ -1,0 +1,11 @@
+<div class="card-header clearfix">
+    <div class="pull-left"><strong>{{title}}</strong></div>
+    <div class="pull-right">
+        <a class="btn btn-link" (click)="onReset()">
+            <i class="fa fa-fw fa-undo"></i> Reset
+        </a>
+        <a class="btn btn-link" (click)="onClose()">
+            <i class="fa fa-fw fa-close"></i> Close
+        </a>
+    </div>
+</div>
