--- src/app/editor/components/StackedItemComponent.html
@@ -1,10 +1,0 @@
-<div class="card card-default" [class.active]="!next">
-    <div class="card-header">Heading</div>
-    <div class="card-body" [id]="contentId">
-        <ng-content></ng-content>
-    </div>
-    <div class="card-footer clearfix">
-        <button class="btn btn-link btn-sm pull-right" (click)="close()">Cancel</button>
-        <button class="btn btn-success btn-sm pull-right" type="submit" (click)="saveAndClose()">Save</button>
-    </div>
-</div>
