+++ src/app/editor/components/iso/gmd/AddressListComponent.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component,ComponentResolver,ViewContainerRef } from '@angular/core';
 import { Address } from '../../../../model/iso/gmd/Address';
 import { AbstractComponent } from '../../AbstractComponent';
 import { AddressComponent } from './AddressComponent';
@@ -14,7 +14,11 @@ class AddressList {
   directives: [AddressComponent]
 })
 export class AddressListComponent extends AbstractComponent<AddressList> {
-
+  constructor(componentResolver:ComponentResolver,
+              viewContainerRef:ViewContainerRef) {
+    super(componentResolver, viewContainerRef);
+  }
+  
   public onAdd() {
     this.model.addresses.push(new Address());
   }
