+++ src/app/editor/components/iso/gmd/OnlineResourceComponent.html
@@ -1,60 +1,35 @@
 <div class="card">
-
-  <card-header title="Online Resource" (reset)="onReset()" [(collapsed)]="collapsed" [(editing)]="editing"></card-header>
-
-  <div [hidden]="collapsed">
-
+    <card-header title="Online Resource" (reset)="onReset()" (close)="close()"></card-header>
     <div class="card-block">
-
-      <div [hidden]="!editing">
         <div class="form-group">
-          <label for="linkage">Linkage</label>
-          <input type="text" class="form-control" [(ngModel)]="model.linkage" required>
+            <label for="linkage">Linkage</label>
+            <input type="text" class="form-control" [(ngModel)]="model.linkage" required>
         </div>
         <div class="form-group">
-          <label for="protocol">Protocol</label>
-          <input type="text" class="form-control" [(ngModel)]="model.protocol">
+            <label for="protocol">Protocol</label>
+            <input type="text" class="form-control" [(ngModel)]="model.protocol">
         </div>
         <div class="form-group">
-          <label for="applicationProfile">Application Profile</label>
-          <input type="text" class="form-control" [(ngModel)]="model.applicationProfile">
+            <label for="applicationProfile">Application Profile</label>
+            <input type="text" class="form-control" [(ngModel)]="model.applicationProfile">
         </div>
         <div class="form-group">
-          <label for="name">Name</label>
-          <input type="text" class="form-control" [(ngModel)]="model.name">
+            <label for="name">Name</label>
+            <input type="text" class="form-control" [(ngModel)]="model.name">
         </div>
         <div class="form-group">
-          <label for="description">Description</label>
-          <input type="text" class="form-control" [(ngModel)]="model.description">
+            <label for="description">Description</label>
+            <input type="text" class="form-control" [(ngModel)]="model.description">
         </div>
         <div class="form-group">
-          <label for="function">Function</label>
-          <select class="form-control" [(ngModel)]="model.function">
-            <option value="download">Download</option>
-            <option value="information">Information</option>
-            <option value="offlineAccess">OfflineAccess</option>
-            <option value="order">Order</option>
-            <option value="search">Search</option>
-          </select>
+            <label for="function">Function</label>
+            <select class="form-control" [(ngModel)]="model.function">
+                <option value="download">Download</option>
+                <option value="information">Information</option>
+                <option value="offlineAccess">OfflineAccess</option>
+                <option value="order">Order</option>
+                <option value="search">Search</option>
+            </select>
         </div>
-      </div>
-
-      <dl class="dl-horizontal" [hidden]="editing">
-        <dt class="col-sm-3 text-truncate">Linkage</dt>
-        <dd class="col-sm-9">{{ model.linkage }}</dd>
-        <dt class="col-sm-3 text-truncate">Protocol</dt>
-        <dd class="col-sm-9">{{ model.protocol }}</dd>
-        <dt class="col-sm-3 text-truncate">Application Profile</dt>
-        <dd class="col-sm-9">{{ model.applicationProfile }}</dd>
-        <dt class="col-sm-3 text-truncate">Name</dt>
-        <dd class="col-sm-9">{{ model.name }}</dd>
-        <dt class="col-sm-3 text-truncate">Description</dt>
-        <dd class="col-sm-9">{{ model.description }}</dd>
-        <dt class="col-sm-3 text-truncate">Function</dt>
-        <dd class="col-sm-9">{{ model.function }}</dd>
-      </dl>
-
     </div>
-
-  </div>
 </div>
