+++ src/app/editor/components/iso/gmd/OnlineResourceComponent.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component,ComponentResolver,ViewContainerRef } from '@angular/core';
 import { OnlineResource } from '../../../../model/iso/gmd/OnlineResource';
 import { AbstractComponent }  from '../../AbstractComponent';
 import { CardHeaderComponent } from '../../CardHeaderComponent';
@@ -9,6 +9,11 @@ import { CardHeaderComponent } from '../../CardHeaderComponent';
 })
 export class OnlineResourceComponent extends AbstractComponent<OnlineResource> {
 
+  constructor(componentResolver:ComponentResolver,
+              viewContainerRef:ViewContainerRef) {
+    super(componentResolver, viewContainerRef);
+  }
+  
   protected createModel(): OnlineResource {
     return new OnlineResource();
   }
