+++ src/app/editor/components/iso/gmd/PhoneComponent.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component,ComponentResolver,ViewContainerRef } from '@angular/core';
 import { Phone } from '../../../../model/iso/gmd/Phone';
 import { AbstractComponent }  from '../../AbstractComponent';
 import { CardHeaderComponent } from '../../CardHeaderComponent';
@@ -11,6 +11,11 @@ import { StringsComponent } from '../../StringsComponent';
 })
 export class PhoneComponent extends AbstractComponent<Phone> {
 
+  constructor(componentResolver:ComponentResolver,
+              viewContainerRef:ViewContainerRef) {
+    super(componentResolver, viewContainerRef);
+  }
+  
   protected createModel(): Phone {
     return new Phone();
   }
